+++ tests/test_iib_client.py
@@ -8,83 +8,93 @@ from iiblib.iib_client import (
     IIBClient,
     IIBException,
 )
-from iiblib.iib_build_details_model import IIBBuildDetailsModel
+from iiblib.iib_build_details_model import IIBBuildDetailsModel, AddModel, RmModel
 from iiblib.iib_build_details_pager import IIBBuildDetailsPager
 
 
 @pytest.fixture
-def fixture_build_details_json():
+def fixture_add_build_details_json():
     json = {
         "id": 1,
+        "arches": ["x86_64"],
         "state": "in_progress",
         "state_reason": "state_reason",
+        "request_type": "add",
         "state_history": [],
-        "from_index": "from_index",
-        "from_index_resolved": "from_index_resolved",
-        "bundles": ["bundles1"],
-        "removed_operators": ["operator1"],
-        "organization": "organization",
+        "batch": 1,
+        "batch_annotations": {"batch_annotations": 1},
+        "logs": {},
+        "updated": "updated",
+        "user": "user@example.com",
         "binary_image": "binary_image",
         "binary_image_resolved": "binary_image_resolved",
-        "index_image": "index_image",
-        "request_type": "request_type",
-        "arches": ["x86_64"],
+        "bundles": ["bundles1"],
         "bundle_mapping": {"bundle_mapping": "map"},
+        "from_index": "from_index",
+        "from_index_resolved": "from_index_resolved",
+        "index_image": "index_image",
+        "removed_operators": ["operator1"],
+        "organization": "organization",
         "omps_operator_version": {"operator": "1.0"},
+        "distribution_scope": "null",
     }
     return json
 
 
 @pytest.fixture
-def fixture_build_details_json2():
+def fixture_rm_build_details_json():
     json = {
         "id": 2,
+        "arches": ["x86_64"],
         "state": "in_progress",
         "state_reason": "state_reason",
+        "request_type": "rm",
         "state_history": [],
-        "from_index": "from_index",
-        "from_index_resolved": "from_index_resolved",
-        "bundles": ["bundles1"],
-        "removed_operators": ["operator1"],
-        "organization": "organization",
+        "batch": 1,
+        "batch_annotations": {"batch_annotations": 1},
+        "logs": {},
+        "updated": "updated",
+        "user": "user@example.com",
         "binary_image": "binary_image",
         "binary_image_resolved": "binary_image_resolved",
-        "index_image": "index_image",
-        "request_type": "request_type",
-        "arches": ["x86_64"],
+        "bundles": ["bundles1"],
         "bundle_mapping": {"bundle_mapping": "map"},
-        "omps_operator_version": {"operator": "1.0"},
-    }
-    return json
-
-
-@pytest.fixture
-def fixture_build_details_json3():
-    json = {
-        "id": 1,
-        "state": "in_progress",
-        "state_reason": "state_reason",
-        "state_history": [],
         "from_index": "from_index",
         "from_index_resolved": "from_index_resolved",
-        "bundles": ["bundles1"],
+        "index_image": "index_image",
         "removed_operators": ["operator1"],
         "organization": "organization",
-        "binary_image": "mapped_binary_image",
-        "binary_image_resolved": "mapped_binary_image_resolved",
-        "index_image": "index_image",
-        "request_type": "request_type",
-        "arches": ["x86_64"],
-        "bundle_mapping": {"bundle_mapping": "map"},
-        "omps_operator_version": {"operator": "1.0"},
+        "distribution_scope": "null",
     }
     return json
 
 
 @pytest.fixture
-def fixture_builds_page1_json(fixture_build_details_json):
+def fixture_regenerate_bundle_build_details_json():
     json = {
-        "items": [fixture_build_details_json],
+        "id": 3,
+        "arches": ["x86_64"],
+        "state": "in_progress",
+        "state_reason": "state_reason",
+        "request_type": "regenerate-bundle",
+        "state_history": [],
+        "batch": 1,
+        "batch_annotations": {"batch_annotations": 1},
+        "logs": {},
+        "updated": "updated",
+        "user": "user@example.com",
+        "bundle_image": "bundle_image",
+        "from_bundle_image": "from_bundle_image",
+        "from_bundle_image_resolved": "from_bundle_image_resolved",
+        "organization": "organization",
+    }
+    return json
+
+
+@pytest.fixture
+def fixture_builds_page1_json(fixture_add_build_details_json):
+    json = {
+        "items": [fixture_add_build_details_json],
         "meta": {
             "first": "",
             "last": "",
@@ -100,9 +110,9 @@ def fixture_builds_page1_json(fixture_build_details_json):
 
 
 @pytest.fixture
-def fixture_builds_page2_json(fixture_build_details_json2):
+def fixture_builds_page2_json(fixture_rm_build_details_json):
     json = {
-        "items": [fixture_build_details_json2],
+        "items": [fixture_rm_build_details_json],
         "meta": {
             "first": "",
             "last": "",
@@ -117,31 +127,58 @@ def fixture_builds_page2_json(fixture_build_details_json2):
     return json
 
 
-def test_iib_client(fixture_build_details_json, fixture_builds_page1_json):
+def test_iib_client(
+    fixture_add_build_details_json,
+    fixture_rm_build_details_json,
+    fixture_regenerate_bundle_build_details_json,
+    fixture_builds_page1_json,
+):
     with requests_mock.Mocker() as m:
         m.register_uri(
             "POST",
             "/api/v1/builds/add",
             status_code=200,
-            json=fixture_build_details_json,
+            json=fixture_add_build_details_json,
         )
         m.register_uri(
             "POST",
             "/api/v1/builds/rm",
             status_code=200,
-            json=fixture_build_details_json,
+            json=fixture_rm_build_details_json,
+        )
+
+        m.register_uri(
+            "POST",
+            "/api/v1/builds/regenerate-bundle",
+            status_code=200,
+            json=fixture_regenerate_bundle_build_details_json,
         )
         m.register_uri(
             "GET", "/api/v1/builds", status_code=200, json=fixture_builds_page1_json
         )
         m.register_uri(
-            "GET", "/api/v1/builds/1", status_code=200, json=fixture_build_details_json
+            "GET",
+            "/api/v1/builds/1",
+            status_code=200,
+            json=fixture_add_build_details_json,
+        )
+        m.register_uri(
+            "GET",
+            "/api/v1/builds/2",
+            status_code=200,
+            json=fixture_rm_build_details_json,
+        )
+        m.register_uri(
+            "GET",
+            "/api/v1/builds/3",
+            status_code=200,
+            json=fixture_regenerate_bundle_build_details_json,
         )
 
         iibc = IIBClient("fake-host")
         assert iibc.add_bundles(
             "index-image", ["bundles-map"], [], binary_image="binary"
-        ) == IIBBuildDetailsModel.from_dict(fixture_build_details_json)
+        ) == AddModel.from_dict(fixture_add_build_details_json)
         assert (
             iibc.add_bundles(
                 "index-image",
@@ -153,13 +190,13 @@ def test_iib_client(fixture_build_details_json, fixture_builds_page1_json):
                 overwrite_from_index=True,
                 overwrite_from_index_token="str",
             )
-            == IIBBuildDetailsModel.from_dict(fixture_build_details_json)
+            == AddModel.from_dict(fixture_add_build_details_json)
         )
         assert (
             iibc.add_bundles(
                 "index-image", ["bundles-map"], [], binary_image="binary", raw=True
             )
-            == fixture_build_details_json
+            == fixture_add_build_details_json
         )
         assert (
             iibc.remove_operators(
@@ -170,18 +207,34 @@ def test_iib_client(fixture_build_details_json, fixture_builds_page1_json):
                 overwrite_from_index=True,
                 overwrite_from_index_token="str",
             )
-            == IIBBuildDetailsModel.from_dict(fixture_build_details_json)
+            == RmModel.from_dict(fixture_rm_build_details_json)
         )
         assert (
             iibc.remove_operators(
                 "index-image", ["operator1"], [], binary_image="binary", raw=True
             )
-            == fixture_build_details_json
+            == fixture_rm_build_details_json
         )
+
+        # get_builds - request_type is "add"
         assert iibc.get_build(1) == IIBBuildDetailsModel.from_dict(
-            fixture_build_details_json
+            fixture_add_build_details_json
         )
-        assert iibc.get_build(1, raw=True) == fixture_build_details_json
+        assert iibc.get_build(1, raw=True) == fixture_add_build_details_json
+
+        # get_builds - request_type is "rm"
+        assert iibc.get_build(2) == IIBBuildDetailsModel.from_dict(
+            fixture_rm_build_details_json
+        )
+        assert iibc.get_build(2, raw=True) == fixture_rm_build_details_json
+
+        # get_builds - request_type is "regenerate-bundle"
+        assert iibc.get_build(3) == IIBBuildDetailsModel.from_dict(
+            fixture_regenerate_bundle_build_details_json
+        )
+        assert (
+            iibc.get_build(3, raw=True) == fixture_regenerate_bundle_build_details_json
+        )
 
         assert iibc.get_builds() == IIBBuildDetailsPager.from_dict(
             iibc, fixture_builds_page1_json
@@ -189,7 +242,9 @@ def test_iib_client(fixture_build_details_json, fixture_builds_page1_json):
         assert iibc.get_builds(raw=True) == fixture_builds_page1_json
 
 
-def test_iib_client_failure(fixture_build_details_json):
+def test_iib_client_no_overwrite_from_index_or_token(
+    fixture_add_build_details_json, fixture_rm_build_details_json
+):
     error_msg = (
         "Either both or neither of overwrite-from-index "
         "and overwrite-from-index-token should be specified."
@@ -199,13 +254,13 @@ def test_iib_client_failure(fixture_build_details_json):
             "POST",
             "/api/v1/builds/add",
             status_code=200,
-            json=fixture_build_details_json,
+            json=fixture_add_build_details_json,
         )
         m.register_uri(
             "POST",
             "/api/v1/builds/rm",
             status_code=200,
-            json=fixture_build_details_json,
+            json=fixture_rm_build_details_json,
         )
         iibc = IIBClient("fake-host")
         with pytest.raises(ValueError, match=error_msg):
@@ -246,71 +301,23 @@ def test_iib_client_failure(fixture_build_details_json):
             )
 
 
-def test_iib_client_no_binary_image(fixture_build_details_json3):
-    with requests_mock.Mocker() as m:
-        m.register_uri(
-            "POST",
-            "/api/v1/builds/add",
-            status_code=200,
-            json=fixture_build_details_json3,
-        )
-        m.register_uri(
-            "POST",
-            "/api/v1/builds/rm",
-            status_code=200,
-            json=fixture_build_details_json3,
-        )
-
-        iibc = IIBClient("fake-host")
-        assert iibc.add_bundles(
-            "index-image", ["bundles-map"], []
-        ) == IIBBuildDetailsModel.from_dict(fixture_build_details_json3)
-        assert (
-            iibc.add_bundles(
-                "index-image",
-                ["bundles-map"],
-                [],
-                cnr_token="cnr",
-                organization="org",
-                overwrite_from_index=True,
-                overwrite_from_index_token="str",
-            )
-            == IIBBuildDetailsModel.from_dict(fixture_build_details_json3)
-        )
-        assert (
-            iibc.add_bundles("index-image", ["bundles-map"], [], raw=True)
-            == fixture_build_details_json3
-        )
-        assert (
-            iibc.remove_operators(
-                "index-image",
-                ["operator1"],
-                [],
-                overwrite_from_index=True,
-                overwrite_from_index_token="str",
-            )
-            == IIBBuildDetailsModel.from_dict(fixture_build_details_json3)
-        )
-        assert (
-            iibc.remove_operators("index-image", ["operator1"], [], raw=True)
-            == fixture_build_details_json3
-        )
-
-
-def test_client_wait_for_build(fixture_build_details_json):
+# there add model used
+def test_client_wait_for_build(fixture_add_build_details_json):
     iibc = IIBClient("fake-host", poll_interval=1)
-    bdetails_finished = copy.copy(fixture_build_details_json)
+    bdetails_finished = copy.copy(fixture_add_build_details_json)
     bdetails_finished["state"] = "complete"
     with requests_mock.Mocker() as m:
         m.register_uri(
             "GET",
             "/api/v1/builds/1",
             [
-                {"json": fixture_build_details_json, "status_code": 200},
+                {"json": fixture_add_build_details_json, "status_code": 200},
                 {"json": bdetails_finished, "status_code": 200},
             ],
         )
-        iibc.wait_for_build(IIBBuildDetailsModel.from_dict(fixture_build_details_json))
+        iibc.wait_for_build(
+            IIBBuildDetailsModel.from_dict(fixture_add_build_details_json)
+        )
 
     bdetails_finished["state"] = "failed"
     with requests_mock.Mocker() as m:
@@ -318,37 +325,43 @@ def test_client_wait_for_build(fixture_build_details_json):
             "GET",
             "/api/v1/builds/1",
             [
-                {"json": fixture_build_details_json, "status_code": 200},
+                {"json": fixture_add_build_details_json, "status_code": 200},
                 {"json": bdetails_finished, "status_code": 200},
             ],
         )
-        iibc.wait_for_build(IIBBuildDetailsModel.from_dict(fixture_build_details_json))
+        iibc.wait_for_build(
+            IIBBuildDetailsModel.from_dict(fixture_add_build_details_json)
+        )
 
 
-def test_client_wait_for_build_retry(fixture_build_details_json):
+# add model used
+def test_client_wait_for_build_retry(fixture_add_build_details_json):
     iibc = IIBClient("fake-host.test", poll_interval=1, retries=10, backoff_factor=0)
 
     with pytest.raises(
         requests.exceptions.RequestException, match=".*Max retries exceeded*."
     ):
-        iibc.wait_for_build(IIBBuildDetailsModel.from_dict(fixture_build_details_json))
+        iibc.wait_for_build(
+            IIBBuildDetailsModel.from_dict(fixture_add_build_details_json)
+        )
 
 
-def test_client_wait_for_build_timeout(fixture_build_details_json):
+# add model used
+def test_client_wait_for_build_timeout(fixture_add_build_details_json):
     # set wait timeout for 2 seconds
     iibc = IIBClient("fake-host.test", poll_interval=1, wait_for_build_timeout=2)
 
     with requests_mock.Mocker() as m:
         m.register_uri(
             "GET",
-            "/api/v1/builds/{}".format(fixture_build_details_json["id"]),
+            "/api/v1/builds/{}".format(fixture_add_build_details_json["id"]),
             status_code=200,
-            json=fixture_build_details_json,
+            json=fixture_add_build_details_json,
         )
 
         with pytest.raises(IIBException, match="Timeout*."):
             iibc.wait_for_build(
-                IIBBuildDetailsModel.from_dict(fixture_build_details_json)
+                IIBBuildDetailsModel.from_dict(fixture_add_build_details_json)
             )